+++ bw2ui/bin/bw2_browser.py
@@ -100,11 +100,14 @@ Working with projects:
 Working with databases:
     ldb: List available databases.
     db name: Go to database name. No quotes needed.
-    s string: Search activity names in current database with string.
-    s -loc {LOCATION} string: Search activity names in current database with \
+    s [string]: Search activity names in current database with string. Without string \
+the search provides no results.
+    s -loc {LOCATION} [string]: Search activity names in current database with \
 string and location LOCATION.
-    s -cat {CAT::SUBCAT::SUBSUBCAT} string: Search activity names in current database \
-with string and category cat, subcat, subcat [useful for biosphere].
+    s -cat {CAT::SUBCAT::SUBSUBCAT} [string]: Search activity names in current \
+database with string and category CAT, SUBCAT, SUBCAT [useful for biosphere].
+    s -rp {REFERENCE PRODUCT} [string]: Search activities in current database that \
+have reference product and optionnaly match string in search.
 
 Working with activities:
     a id: Go to activity id in current database. Complex ids in quotes.
@@ -1071,8 +1074,13 @@ Autosave is turned %(autosave)s.""" % {
             elif "-cat" in arg:
                 re1a = r"(-cat\s)"  # Any Single Whitespace Character 1
                 search_criterion = "category"
+            elif "-rp" in arg:
+                re1a = r"(-rp\s)"  # Any Single Whitespace Character 1
+                search_criterion = "reference product"
             re1b = r"(\{.*\})"  # Curly Braces 1
-            re2 = r"\s(.+)"
+            re2 = (
+                r"(?:\s(.+))?"  # at least a space, and then 1 to n chars, but optional
+            )
             rg = re.compile(re1a + re1b + re2, re.IGNORECASE | re.DOTALL)
             m = rg.search(arg)
             needle = arg  # Find the needle in the haystack
@@ -1081,6 +1089,12 @@ Autosave is turned %(autosave)s.""" % {
                 return
             elif m is None and "-cat " in arg:
                 print("Missing category in curly braces in command: -cat {water} ...")
+                return
+            elif m is None and "-rp" in arg:
+                print(
+                    "Missing reference product in curly braces in command: -rp "
+                    "{electricity high voltage} ..."
+                )
                 return
             elif m:
                 c2 = m.group(2)
@@ -1094,25 +1108,49 @@ Autosave is turned %(autosave)s.""" % {
 
             specials = [" ", "/", "-", "&"]
             if m and search_criterion == "location":
-                if not any(
-                    special in criterion_value for special in specials
-                ):  # complex criterion_value
+                if not any(special in criterion_value for special in specials):
                     criteria = {"location": criterion_value}
-                    results = Database(self.database).search(
-                        needle, filter=criteria, limit=self.search_limit
-                    )
-                else:
-                    results = Database(self.database).search(needle, limit=None)
+                    if needle:
+                        results = Database(self.database).search(
+                            needle, filter=criteria, limit=self.search_limit
+                        )
+                    else:
+                        results = [
+                            a
+                            for a in Database(self.database)
+                            if a.get("location", "").casefold()
+                            == criterion_value.casefold()
+                        ]
+
+                else:  # complex criterion_value
+                    if needle:
+                        results = Database(self.database).search(needle, limit=None)
+                    else:
+                        results = Database(self.database)
                     results = [
                         r
                         for r in results
-                        if r[search_criterion].lower() == criterion_value.lower()
+                        if r[search_criterion].casefold() == criterion_value.casefold()
                     ]
             elif m and search_criterion == "category":
                 criteria = {"categories": criterion_value.split("::")}
                 results = Database(self.database).search(
                     needle, filter=criteria, limit=self.search_limit
                 )
+            elif m and search_criterion == "reference product":
+                if needle is None:
+                    results = [
+                        a
+                        for a in Database(self.database)
+                        if a[search_criterion].casefold() == criterion_value.casefold()
+                    ]
+                else:
+                    criteria = {"product": criterion_value}
+                    results = Database(self.database).search(
+                        needle,
+                        filter=criteria,
+                        limit=self.search_limit,
+                    )
             else:
                 results = Database(self.database).search(
                     needle, limit=self.search_limit